+++ packages/core/src/utils/ConstantsUtil.ts
@@ -35,9 +35,6 @@ export const ConstantsUtil = {
     'ASIA/HARBIN'
   ],
 
-  CONNECTOR_RDNS_MAP: {
-    coinbaseWallet: 'com.coinbase.wallet'
-  } as Record<string, string>,
   /**
    * Network name to Coinbase Pay SDK chain name map object
    * @see supported chain names on Coinbase for Pay SDK: https://github.com/coinbase/cbpay-js/blob/d4bda2c05c4d5917c8db6a05476b603546046394/src/types/onramp.ts