+++ packages/ethers/src/client.ts
@@ -42,6 +42,7 @@ import { W3mFrameProvider, W3mFrameHelpers, W3mFrameRpcConstants } from '@web3mo
 import type { CombinedProvider } from '@web3modal/scaffold-utils/ethers'
 import { BrowserProvider } from 'ethers'
 import { JsonRpcSigner } from 'ethers'
+import { NetworkUtil } from '@web3modal/common'
 
 // -- Types ---------------------------------------------------------------------
 export interface Web3ModalClientOptions extends Omit<LibraryOptions, 'defaultChain' | 'tokens'> {
@@ -52,7 +53,6 @@ export interface Web3ModalClientOptions extends Omit<LibraryOptions, 'defaultCha
   chainImages?: Record<number, string>
   connectorImages?: Record<string, string>
   tokens?: Record<number, Token>
-  enableSmartAccounts?: boolean
 }
 
 export type Web3ModalOptions = Omit<Web3ModalClientOptions, '_sdkVersion'>
@@ -75,13 +75,8 @@ interface Info {
   rdns: string
 }
 
-interface Wallet {
+interface EIP6963ProviderDetail {
   info: Info
-  provider: Provider
-}
-
-interface IEIP6963Provider {
-  name: string
   provider: Provider
 }
 
@@ -93,7 +88,7 @@ interface ExternalProvider extends EthereumProvider {
 export class Web3Modal extends Web3ModalScaffold {
   private hasSyncedConnectedAccount = false
 
-  private EIP6963Providers: IEIP6963Provider[] = []
+  private EIP6963Providers: EIP6963ProviderDetail[] = []
 
   private walletConnectProvider?: EthereumProvider
 
@@ -131,7 +126,7 @@ export class Web3Modal extends Web3ModalScaffold {
 
     const networkControllerClient: NetworkControllerClient = {
       switchCaipNetwork: async caipNetwork => {
-        const chainId = HelpersUtil.caipNetworkIdToNumber(caipNetwork?.id)
+        const chainId = NetworkUtil.caipNetworkIdToNumber(caipNetwork?.id)
         if (chainId) {
           try {
             EthersStoreUtil.setError(undefined)
@@ -405,7 +400,7 @@ export class Web3Modal extends Web3ModalScaffold {
 
     return {
       ...state,
-      selectedNetworkId: HelpersUtil.caipNetworkIdToNumber(state.selectedNetworkId)
+      selectedNetworkId: NetworkUtil.caipNetworkIdToNumber(state.selectedNetworkId)
     }
   }
 
@@ -414,7 +409,7 @@ export class Web3Modal extends Web3ModalScaffold {
     return super.subscribeState(state =>
       callback({
         ...state,
-        selectedNetworkId: HelpersUtil.caipNetworkIdToNumber(state.selectedNetworkId)
+        selectedNetworkId: NetworkUtil.caipNetworkIdToNumber(state.selectedNetworkId)
       })
     )
   }
@@ -582,10 +577,10 @@ export class Web3Modal extends Web3ModalScaffold {
     const currentActiveWallet = window?.localStorage.getItem(EthersConstantsUtil.WALLET_ID)
     if (currentActiveWallet) {
       const currentProvider = this.EIP6963Providers.find(
-        provider => provider.name === currentActiveWallet
+        provider => provider.info.name === currentActiveWallet
       )
       if (currentProvider) {
-        this.setEIP6963Provider(currentProvider.provider, currentProvider.name)
+        this.setEIP6963Provider(currentProvider.provider, currentProvider.info.name)
       }
     }
   }
@@ -656,40 +651,28 @@ export class Web3Modal extends Web3ModalScaffold {
     }
   }
 
-  private async initSmartAccount(
-    chainId: number
-  ): Promise<{ isDeployed: boolean; address?: string }> {
-    if (!this.emailProvider || !this.options?.enableSmartAccounts) {
-      return { isDeployed: false }
-    }
-    const { smartAccountEnabledNetworks } =
-      await this.emailProvider.getSmartAccountEnabledNetworks()
-
-    if (!smartAccountEnabledNetworks.includes(chainId)) {
-      return { isDeployed: false }
-    }
-
-    return await this.emailProvider.initSmartAccount()
-  }
-
   private async setEmailProvider() {
     window?.localStorage.setItem(EthersConstantsUtil.WALLET_ID, ConstantsUtil.EMAIL_CONNECTOR_ID)
 
     if (this.emailProvider) {
-      const { address, chainId } = await this.emailProvider.connect()
+      const preferredAccountType = W3mFrameHelpers.getPreferredAccountType()
+      const [{ address, chainId, smartAccountDeployed }, { smartAccountEnabledNetworks }] =
+        await Promise.all([
+          this.emailProvider.connect({
+            chainId: this.getChainId(),
+            preferredAccountType
+          }),
+          this.emailProvider.getSmartAccountEnabledNetworks()
+        ])
       super.setLoading(false)
       if (address && chainId) {
         EthersStoreUtil.setChainId(chainId)
         EthersStoreUtil.setProviderType(ConstantsUtil.EMAIL_CONNECTOR_ID as 'w3mEmail')
         EthersStoreUtil.setProvider(this.emailProvider as unknown as CombinedProvider)
         EthersStoreUtil.setIsConnected(true)
-        const { isDeployed, address: smartAccountAddress } = await this.initSmartAccount(chainId)
-        this.setSmartAccountDeployed(isDeployed)
-        if (isDeployed && smartAccountAddress) {
-          EthersStoreUtil.setAddress(smartAccountAddress as Address)
-        } else {
-          EthersStoreUtil.setAddress(address as Address)
-        }
+        EthersStoreUtil.setAddress(address as Address)
+        this.setSmartAccountDeployed(Boolean(smartAccountDeployed))
+        this.setSmartAccountEnabledNetworks(smartAccountEnabledNetworks)
 
         this.watchEmail()
         this.watchModal()
@@ -869,7 +852,19 @@ export class Web3Modal extends Web3ModalScaffold {
         super.setLoading(false)
       })
       this.emailProvider.onIsConnected(() => {
+        this.setIsConnected(true)
         super.setLoading(false)
+      })
+
+      this.emailProvider.onSetPreferredAccount(({ address }) => {
+        if (!address) {
+          return
+        }
+        const chainId = NetworkUtil.caipNetworkIdToNumber(this.getCaipNetwork()?.id)
+        EthersStoreUtil.setAddress(address as Address)
+        EthersStoreUtil.setChainId(chainId)
+        EthersStoreUtil.setIsConnected(true)
+        this.syncAccount()
       })
     }
   }
@@ -897,7 +892,7 @@ export class Web3Modal extends Web3ModalScaffold {
       this.setIsConnected(isConnected)
 
       this.setCaipAddress(caipAddress)
-
+      this.syncConnectedWalletInfo()
       await Promise.all([
         this.syncProfile(address),
         this.syncBalance(address),
@@ -987,6 +982,39 @@ export class Web3Modal extends Web3ModalScaffold {
           this.setBalance(formattedBalance, chain.currency)
         }
       }
+    }
+  }
+
+  private syncConnectedWalletInfo() {
+    const currentActiveWallet = window?.localStorage.getItem(EthersConstantsUtil.WALLET_ID)
+    const providerType = EthersStoreUtil.state.providerType
+
+    if (providerType === ConstantsUtil.EIP6963_CONNECTOR_ID) {
+      if (currentActiveWallet) {
+        const currentProvider = this.EIP6963Providers.find(
+          provider => provider.info.name === currentActiveWallet
+        )
+
+        if (currentProvider) {
+          this.setConnectedWalletInfo({
+            ...currentProvider.info
+          })
+        }
+      }
+    } else if (providerType === ConstantsUtil.WALLET_CONNECT_CONNECTOR_ID) {
+      const provider = EthersStoreUtil.state.provider as unknown as EthereumProvider
+
+      if (provider.session) {
+        this.setConnectedWalletInfo({
+          ...provider.session.peer.metadata,
+          name: provider.session.peer.metadata.name,
+          icon: provider.session.peer.metadata.icons?.[0]
+        })
+      }
+    } else if (currentActiveWallet) {
+      this.setConnectedWalletInfo({
+        name: currentActiveWallet
+      })
     }
   }
 
@@ -1095,8 +1123,18 @@ export class Web3Modal extends Web3ModalScaffold {
       } else if (providerType === ConstantsUtil.EMAIL_CONNECTOR_ID) {
         if (this.emailProvider && chain?.chainId) {
           try {
-            await this.emailProvider?.switchNetwork(chain?.chainId)
+            await this.emailProvider.switchNetwork(chain?.chainId)
             EthersStoreUtil.setChainId(chain.chainId)
+
+            this.emailProvider
+              .connect({
+                chainId: chain?.chainId,
+                preferredAccountType: W3mFrameHelpers.getPreferredAccountType()
+              })
+              .then(({ address }) => {
+                EthersStoreUtil.setAddress(address as Address)
+                this.syncAccount()
+              })
           } catch {
             throw new Error('Switching chain failed')
           }
@@ -1173,7 +1211,7 @@ export class Web3Modal extends Web3ModalScaffold {
     }
   }
 
-  private eip6963EventHandler(event: CustomEventInit<Wallet>) {
+  private eip6963EventHandler(event: CustomEventInit<EIP6963ProviderDetail>) {
     if (event.detail) {
       const { info, provider } = event.detail
       const connectors = this.getConnectors()
@@ -1193,8 +1231,8 @@ export class Web3Modal extends Web3ModalScaffold {
           })
 
           const eip6963ProviderObj = {
-            name: info.name,
-            provider
+            provider,
+            info
           }
 
           this.EIP6963Providers.push(eip6963ProviderObj)