+++ packages/ui/src/composites/wui-input-numeric/styles.ts
@@ -19,7 +19,11 @@ export default css`
     text-align: center;
     color: var(--wui-color-fg-100);
     caret-color: var(--wui-color-accent-100);
-    transition: all var(--wui-ease-inout-power-1) var(--wui-duration-lg);
+    transition:
+      background-color var(--wui-ease-inout-power-1) var(--wui-duration-md),
+      border-color var(--wui-ease-inout-power-1) var(--wui-duration-md),
+      box-shadow var(--wui-ease-inout-power-1) var(--wui-duration-md);
+    will-change: background-color, border-color, box-shadow;
     box-sizing: border-box;
     -webkit-appearance: none;
     -moz-appearance: textfield;
@@ -43,13 +47,13 @@ export default css`
   }
 
   input:focus:enabled {
-    transition: all var(--wui-ease-out-power-2) var(--wui-duration-sm);
     background-color: var(--wui-gray-glass-015);
     border: 1px solid var(--wui-color-accent-100);
     -webkit-box-shadow: 0px 0px 0px 4px var(--wui-box-shadow-blue);
     -moz-box-shadow: 0px 0px 0px 4px var(--wui-box-shadow-blue);
     box-shadow: 0px 0px 0px 4px var(--wui-box-shadow-blue);
   }
+
   @media (hover: hover) and (pointer: fine) {
     input:hover:enabled {
       background-color: var(--wui-gray-glass-015);