+++ packages/ui/src/utils/TypeUtil.ts
@@ -168,7 +168,9 @@ export type VisualType =
   | 'defi'
   | 'defiAlt'
   | 'eth'
+  | 'google'
   | 'layers'
+  | 'lightbulb'
   | 'lock'
   | 'login'
   | 'network'
@@ -180,6 +182,7 @@ export type VisualType =
   | 'moonpay'
   | 'stripe'
   | 'paypal'
+  | 'pencil'
 
 export type VisualSize = 'sm' | 'md' | 'lg'