+++ examples/next-wagmi/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @examples/next-wagmi
 
+## 4.1.3-5f2ae345.1
+
+### Patch Changes
+
+- canary: test imports
+
+- Updated dependencies []:
+  - @web3modal/wagmi@4.1.3-5f2ae345.1
+
+## 4.1.3-8e039e.0
+
+### Patch Changes
+
+- feat: update optional dependencies
+
+- Updated dependencies []:
+  - @web3modal/wagmi@4.1.3-8e039e.0
+
 ## 4.1.2
 
 ### Patch Changes